funciones y arrays W3Schools

+++ W3Schools/17 - PHP Functions/demo09.php
@@ -5,8 +5,8 @@
 
 <?php
     // PHP Return Type Declarations
-    function addNumbers(float $a, float $b) : float {
-        return $a + $b;
+    function addNumbers(float $a, float $b) : int {
+        return (int)($a + $b);
     }
     echo addNumbers(1.2, 5.2);
 ?>
